Bugfix for type in ioctl wrapper

+++ aiscm/v4l2.c
@@ -41,7 +41,7 @@ struct videodev2_t {
   char capture;
 };
 
-static int xioctl(int fd, int request, void *arg)
+static int xioctl(int fd, unsigned long int request, void *arg)
 {
   int r;
   do {
